pyproject.toml
[project]
name = "chomp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.3.5",
    "pexpect>=4.9.0",
    "pillow>=11.0.0",
]
